:# Commit Convention ♻️ refactor : purpose enum 위치 users.utils로 이동
